requirements.txt
requests
urllib3
pysocks
